+++ packages/sanity/src/core/form/members/object/fields/PrimitiveField.tsx
@@ -126,6 +126,7 @@ export function PrimitiveField(props: {
   const inputProps = useMemo((): Omit<PrimitiveInputProps, 'renderDefault'> => {
     return {
       value: member.field.value as any,
+      __unstable_computeDiff: member.field.__unstable_computeDiff,
       readOnly: member.field.readOnly,
       schemaType: member.field.schemaType as any,
       changed: member.field.changed,
@@ -141,6 +142,7 @@ export function PrimitiveField(props: {
     }
   }, [
     member.field.value,
+    member.field.__unstable_computeDiff,
     member.field.readOnly,
     member.field.schemaType,
     member.field.changed,