+++ packages/sanity/src/core/form/store/types/nodes.ts
@@ -1,3 +1,4 @@
+import {type Diff} from '@sanity/diff'
 import {
   type ArraySchemaType,
   type BooleanSchemaType,
@@ -11,14 +12,43 @@ import {
 
 import {type FormNodePresence} from '../../../presence'
 import {type ObjectItem} from '../../types'
+import {type ProvenanceDiffAnnotation} from './diff'
 import {type FormFieldGroup} from './fieldGroup'
 import {type ArrayOfObjectsMember, type ArrayOfPrimitivesMember, type ObjectMember} from './members'
+
+/**
+ * @public
+ */
+export type ComputeDiff<Annotation> = (value: unknown) => Diff<Annotation>
+
+/**
+ * Props that encapsulate changes in the node's value.
+ *
+ * @public
+ */
+export interface NodeDiffProps<Annotation> {
+  /**
+   * A function that takes any value and produces a diff between that value and the value the node
+   * is being compared to.
+   *
+   * This can be used to compute a diff optimistically.
+   *
+   * This is marked as unstable because the API may need to evolve as we iterate on the advanced
+   * version control functionality. It will be stabilised when that project has matured.
+   */
+  __unstable_computeDiff: ComputeDiff<Annotation>
+  /**
+   * Whether the current value is different to the value the node is being compared to.
+   */
+  changed: boolean
+}
 
 /**
  * @hidden
  * @public
  */
-export interface BaseFormNode<T = unknown, S extends SchemaType = SchemaType> {
+export interface BaseFormNode<T = unknown, S extends SchemaType = SchemaType>
+  extends NodeDiffProps<ProvenanceDiffAnnotation> {
   // constants
   /** The unique identifier of the node. */
   id: string
@@ -42,8 +72,6 @@ export interface BaseFormNode<T = unknown, S extends SchemaType = SchemaType> {
   readOnly?: boolean
   /** Whether the node is focused. */
   focused?: boolean
-  /** Whether the node has changes in a draft. */
-  changed: boolean
 }
 
 /** @internal */